requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
croniter==6.0.0
distro==1.9.0
dnspython==2.8.0
fastapi==0.116.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
multidict==6.6.4
mypy_extensions==1.1.0
openai==1.107.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
